chore: consolidate core tests
