+++ core/tests/test_comment_reply.py
@@ -6,7 +6,7 @@ from django.test import TestCase
 from django.urls import reverse
 from django.utils import timezone
 
-from .models import Comment, Community, Post
+from core.models import Comment, Community, Post
 
 
 class CommentReplyTests(TestCase):
