+++ core/tests/test_deleted_post_visibility.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.urls import reverse
 
-from .models import Community, Post
+from core.models import Community, Post
 
 
 class DeletedPostVisibilityTests(TestCase):
