+++ core/tests/test_feed_tabs.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.urls import reverse
 
-from .models import Community, Post
+from core.models import Community, Post
 
 
 class FeedTabOrderTests(TestCase):
