+++ core/tests/test_mod_actions.py
@@ -3,7 +3,7 @@ from django.core.cache import cache
 from django.test import TestCase
 from django.urls import reverse
 
-from .models import Comment, Community, Post
+from core.models import Comment, Community, Post
 
 
 class ModActionTests(TestCase):
