+++ core/tests/test_transparency_posts.py
@@ -4,8 +4,8 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase, override_settings
 from django.urls import reverse
 
-from .models import Community, Post
-from .votes import apply_vote
+from core.models import Community, Post
+from core.votes import apply_vote
 
 
 @override_settings(ASTRO_EARLY_VOTES_N=5, ASTRO_MIN_EARLY_VOTES=3, ASTRO_EARLY_SHARE_RED=0.5)
