+++ core/tests/test_user_profile.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.test import TestCase
 from django.urls import reverse
 
-from .models import Comment, Community, Post
+from core.models import Comment, Community, Post
 
 
 class UserProfileTests(TestCase):
