+++ core/tests/test_vote_model.py
@@ -2,7 +2,7 @@ from django.contrib.auth import get_user_model
 from django.db import IntegrityError
 from django.test import TestCase
 
-from .models import Community, Post, Comment, Vote
+from core.models import Community, Post, Comment, Vote
 
 
 class VoteModelTests(TestCase):
